Corrected Hindi version of "Hello!"
Hello Purple Scientist,

I'm TuxedLinux, this is my newer account. You probably already know about this.

My mother tongue is not Hindi (we don't speak Hindi in the south of India) but I can understand it a bit. हैलो is literally pronounced as Hello, the correct Hindi version is नमस्ते, pronounced as Namaste.

Not sure if you'll ever see this though, you've been offline for an long time.